only build packages if new version

+++ bin/update-to-latest.py
@@ -8,6 +8,7 @@ import sys
 import re
 from lxml import etree
 import wmi
+import glob
 
 logging.basicConfig(level=logging.DEBUG)
 
@@ -24,7 +25,7 @@ parser.add_argument('--bundle_dir', '-b', nargs='?', default = def_bdir,
                     help='directory in which to make fake bundles')
 args = parser.parse_args()
 
-# get version
+# get git version
 gitver = subprocess.check_output(
     ['git', '--git-dir', os.path.join(args.gitdir, '.git'),
      'describe', '--tags', '--long' ],
@@ -35,56 +36,74 @@ vlist = gitver.split('-')
 vlist.pop()
 commits = vlist.pop()
 version = '{}.{}'.format('-'.join(vlist), commits)
+logging.info('Git version {}'.format(version))
 
-# Clean dist
 distdir = os.path.join(args.gitdir, 'dist')
+# get packaged version
+pkg_version = None
 for f in os.listdir(distdir):
     fpath = os.path.join(distdir, f)
-    if os.path.isfile(fpath) and f.startswith('machination-client'):
-        logging.info('cleaning pkg {}'.format(f))
-#        os.remove(fpath)
+    if os.path.isfile(fpath):
+        m = re.match(r'machination-client-core-(\d+\.\d+\.\d+)', f)
+        if m:
+            pkg_version = m.group(1)
+            break
+logging.info('Packaged version {}'.format(pkg_version))
 
-# Clean bundles
-for d in os.listdir(args.bundle_dir):
-    dpath = os.path.join(args.bundle_dir, d)
-    if os.path.isdir(dpath) and d. startswith('machination-client'):
-        logging.info('cleaning bundle {}'.format(d))
-        shutil.rmtree(dpath)
+if version != pkg_version:
+    # Make new packages
 
-# Make new packages
-#subprocess.check_call(
-#    [sys.executable,
-#     os.path.join(args.gitdir, 'setup.py'),
-#     bdist_cmd]
-#    )
+    # Clean dist
+    for f in os.listdir(distdir):
+        fpath = os.path.join(distdir, f)
+        if os.path.isfile(fpath) and f.startswith('machination-client'):
+            logging.info('cleaning pkg {}'.format(f))
+            os.remove(fpath)
 
-# Clean .egg_info dirs
-for d in os.listdir(args.gitdir):
-    dpath = os.path.join(args.gitdir, d)
-    if os.path.isdir(dpath) and d. endswith('.egg-info'):
-        logging.info('cleaning egg-info {}'.format(d))
-        shutil.rmtree(dpath)
+    # Clean bundles
+    for d in os.listdir(args.bundle_dir):
+        dpath = os.path.join(args.bundle_dir, d)
+        if os.path.isdir(dpath) and d. startswith('machination-client'):
+            logging.info('cleaning bundle {}'.format(d))
+            shutil.rmtree(dpath)
+
+    # Make new packages
+    subprocess.check_call(
+        [sys.executable,
+         os.path.join(args.gitdir, 'setup.py'),
+         bdist_cmd]
+        )
+
+    # Clean .egg_info dirs
+    for d in os.listdir(args.gitdir):
+        dpath = os.path.join(args.gitdir, d)
+        if os.path.isdir(dpath) and d. endswith('.egg-info'):
+            logging.info('cleaning egg-info {}'.format(d))
+            shutil.rmtree(dpath)
 
 # Make bundles
 pkgids = []
 for f in os.listdir(distdir):
     fpath = os.path.join(distdir, f)
-#    logging.debug('testing {} for pkgid'.format(f))
+    logging.debug('testing {} for pkgid'.format(f))
     m = re.match(r'(machination-client-.*)-{}'.format(version), f)
     pkgname = m.group(1)
     logging.info('bundling {} version {}'.format(pkgname, version))
     pkgid = '{}-{}'.format(pkgname, version)
     pkgids.append(pkgid)
-    os.mkdir(os.path.join(args.bundle_dir, pkgid))
-    os.mkdir(os.path.join(args.bundle_dir, pkgid, 'files'))
-    shutil.copy(fpath, os.path.join(args.bundle_dir, pkgid, 'files'))
+    if not os.path.exists(os.path.join(args.bundle_dir, pkgid)):
+        os.mkdir(os.path.join(args.bundle_dir, pkgid))
+        os.mkdir(os.path.join(args.bundle_dir, pkgid, 'files'))
+        shutil.copy(fpath, os.path.join(args.bundle_dir, pkgid, 'files'))
 
 # Construct installed_version.xml
-iv = etree.Element('iv', budleDir = args.bundle_dir)
+logging.info('constructing iv.xml with bundleDir {}'.format(args.bundle_dir))
+iv = etree.Element('iv', bundleDir = args.bundle_dir)
 current = etree.Element('installedVersion', version='current')
 desired = etree.Element('installedVersion', version='desired')
 # desired from pkgids
 for pkgid in pkgids:
+    logging.info('adding {} to desired'.format(pkgid))
     desired.append(
         etree.Element('machinationFetcherBundle', id = pkgid)
         )
@@ -92,18 +111,22 @@ for pkgid in pkgids:
 # TODO(colin): support other OSes
 con = wmi.WMI()
 for prod in con.query(
-    "select * from Win32Product where Name like 'Python machination-client%'"
+    "select * from Win32_Product where Name like 'Python machination-client%'"
     ):
+    logging.info('adding {} to current'.format(pkgid))
     current.append(etree.Element('machinationFetcherBundle', id=prod.Name[7:]))
 iv.append(current)
 iv.append(desired)
 # write to file
 with open('iv.xml', "w") as ivf:
-    ivf.write(etree.tostring(iv))
+    ivf.write(etree.tostring(iv, pretty_print=True).decode())
 
 # invoke self update
-os.execl(
+args = [
     sys.executable,
+    'machination-self-update',
     os.path.join(args.gitdir, 'bin', 'machination-self-update.py'),
     'iv.xml'
-    )
+    ]
+logging.debug(args)
+os.execl(*args)
